move Inspector equalTo implementation and test
